moved maps to separate folder - segregating by functionality
