+++ app/commandHandlersFactory.js
@@ -8,8 +8,8 @@ import leaveLadderHandler from './handlers/leaveLadderHandler';
 import addResultHandler from './handlers/addResultHandler';
 import showStatsHandler from './handlers/showStatsHandler';
 import rankingHandler from './handlers/rankingHandler';
-import {mapUpVoteHandler, mapDownVoteHandler} from './handlers/mapVoteHandlers';
-import listMapsHandler from './handlers/listMapsHandler';
+import {mapUpVoteHandler, mapDownVoteHandler} from './maps/mapVoteHandlers';
+import listMapsHandler from './maps/listMapsHandler';
 import showLaddersHandler from './handlers/showLaddersHandler';
 import thisIsNotTheCommandYouAreLookingFor from './handlers/nullHandler';
 import validateLadderExistenceDecorator from './validation/validateLadderExistenceDecorator.js';
