+++ app/maps/mapVoteHandlers.js
